spotforecast is a core class
